Dynamic Pagination of Issues & Pull Requests (#9)
## What's changing?

This refactors the data table implementations to add support for a
lazily loaded data table implementation. As implemented, if the user
scrolls within 5 rows of the bottom of the table, it will trigger an
attempt to load more data from the specified load function. If no data
is received from that function, then it will set a flag indicating that
no additional data can be loaded.

## Why is it changing?

This allows users to browse PRs and Issues on a larger repo without hard
limits on the record count.

---------

Co-authored-by: gizmo385 <gizmo385@users.noreply.github.com>