Universally set keys for all data table rows
This commit changes the interfaces for some of the DataTable
abstractions to ensure that we always set a key for each row in the
table. This allows us to more easily update entries in the table,
prevent duplicates, and generally cleans up the interface.

Additionally, this'll hopefully allow us to add heavier local caching to
improve cold start times by preloading the most recent request results
into the UI on startup.