Better offline non-HTTP caching (#33)
This PR overhauls the searchable data table interface to remove a number
of repeated patterns throughout the codebase. The new interfaces include
caching and handles all of the type conversions into the data table
format for our base models.

This makes the cold start experience better since it'll load with a
local file cache more often, which was a long standing desire from #2.

---------

Co-authored-by: gizmo385 <gizmo385@users.noreply.github.com>