Fix race condition in cache loading
This most commonly affected small lists but when cached data was being
loaded after swapping repos in the UI, it was possible for a
sufficiently small list to trigger an attempt to load more data. Since
Since the load function wasn't previously cleared when rows were
cleared, this could result in data from the old repo being loaded into
the cache for the newly selected repo.