Remove need for @classproperty
These were previously used to provide a property-like behavior for class
attributes on LazyGithubContext and LazyGithubBindings. Unfortunately,
making class properties type check is very annoying and LSPs have
trouble understanding them, which made interacting with the context
object annoying.

As an alternative, I've renamed LazyGithubContext to _LazyGithubContext
and defined a top-level variable in the context module named
LazyGithubContext which is an instance of _LazyGithubContext. Alongside
this, I've changed all usages of @classproperty to @property, which
allows for proper type checking.

This commit includes a number of fixes for typing errors that this was
previously hiding.